% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{COMMUNES}
\title{Données contenues dans le package}
\format{\code{sf}}
\usage{
COMMUNES
}
\description{
La couche cartographique simplifiée des communes de l'année "annee_ref"
est contenue dans ce package.
}
\keyword{datasets}
