% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deplacer_DOM.R
\encoding{UTF-8}
\name{deplacer_DOM}
\alias{deplacer_DOM}
\title{Modifier une couche cartographique de type sf en indiquant la position souhaitée des DOM}
\usage{
deplacer_DOM(
  objet,
  nivsupra = colnames(objet)[1],
  positions_type = c("topleft"),
  positions = NULL,
  rotations = rep(0, 5),
  zooms = rep(1, 5),
  projection = st_crs(objet)$proj4string
)
}
\arguments{
\item{objet}{est un fond de carte de la France (métropolitaine et 5 DOM) de type sf. Les niveaux géographiques possibles sont ceux du package CARTElette c'est-à-dire :
- "DEP" : départements
- "REG" : régions
- "EPCI" : EPCI au 01/01/20XX
- "ARR" : arrondissements au 01/01/20XX
- "CV" : cantons-villes au 01/01/20XX
- "ZE2010" : zones d'emploi 2010 (avant 2020)
- "ZE2020" : zones d'emploi 2020 (après 2020)
- "UU2010" : unités urbaines 2010 (avant 2020)
- "UU2020" : unités urbaines 2020 (après 2020)
- "AU2010" : aires urbaines 2010 (avant 2020)
- "AAV2020" : aires d'attraction des villes 2020 (après 2020)
- "BV2012" : bassins de vie 2012}

\item{nivsupra}{est une chaîne de caractères qui indique le nom du niveau supra-communal concerné. Cette chaîne de caractère doit également correspondre à la colonne de l'objet sf contenant les codes géographiques. Il peut s'agir de :
- "DEP" : départements
- "REG" : régions
- "EPCI" : EPCI au 01/01/20XX
- "ARR" : arrondissements au 01/01/20XX
- "CV" : cantons-villes au 01/01/20XX
- "ZE2010" : zones d'emploi 2010 (avant 2020)
- "ZE2020" : zones d'emploi 2020 (après 2020)
- "UU2010" : unités urbaines 2010 (avant 2020)
- "UU2020" : unités urbaines 2020 (après 2020)
- "AU2010" : aires urbaines 2010 (avant 2020)
- "AAV2020" : aires d'attraction des villes 2020 (après 2020)
- "BV2012" : bassins de vie 2012}

\item{positions_type}{est un type de position proposé parmi les positions par défaut :
- "topleft" : DOM positionnés en haut à gauche}

\item{positions}{est une liste de positions pour les DOM (à indiquer dans l'ordre :  Guadeloupe, Martinique, Guyane, Réunion et Mayotte) . Il peut s'agir par exemple de résultats des fonctions positionner_DOM_grille ou encore positionner_DOM_glisser. Par défaut vaut NULL (cf. positions_type)}

\item{rotations}{vecteur de 5 angles au format trigonométrique (pi/2 pour 90°...) qui correspondent aux rotations souhaitées pour les DOM (à indiquer dans l'ordre :  Guadeloupe, Martinique, Guyane, Réunion et Mayotte). Par défaut vaut rep(0,5) (pas de rotation).}

\item{zooms}{vecteur de 5 zooms qui correspondent aux grossissements souhaités pour les DOM (à indiquer dans l'ordre :  Guadeloupe, Martinique, Guyane, Réunion et Mayotte). Par défaut vaut rep(1,5) (pas de zoom).}

\item{projection}{projection de la couche cartographique à modifier. vaut NULL si la projection souhaitée est WGS84 ("+proj=longlat +datum=WGS84 +no_defs"). Sinon peut être devinée en faisant sf::st_crs(couche_sf)$proj4string}
}
\description{
Modifier une couche cartographique de type sf en indiquant la position souhaitée des DOM
}
\details{
La fonction renvoie un objet de type sf
}
\examples{
\dontrun{
# Exemple 1 : choisir precisement la position des DOM
ze <- charger_carte(COG = 2021, nivsupra = "ZE2020")
#positions <- positionner_DOM_glisser(objet=ze)
#positions <- positionner_DOM_grille(projection = st_crs(ze)$proj4string)
positions <- list(c(-5.074931, 46.920490), c(-6.768008, 49.571175), c(-2.65836, 45.08238),
                 c( -6.195586, 45.084970), c(4.948049, 41.193759))
ze_final <- deplacer_DOM(objet = ze, positions = positions)
par(mar=c(0, 0, 0, 0))
plot(sf::st_geometry(ze_final))

# Exemple 2 : choisir des positions de DOM par defaut
ze_final_2 <- deplacer_DOM(objet = ze, positions_type = "topleft")
par(mar = c(0, 0, 0, 0))
plot(sf::st_geometry(ze_final_2))

# Exemple 3 : changer les zooms des DOM et les faire pivoter
ze_final_3 <- deplacer_DOM(objet = ze, positions_type = "topleft",
rotations = c(90,180,90,-90,0), zooms = c(1,1,2,0.5,1))
par(mar = c(0,0,0,0))
plot(sf::st_geometry(ze_final_3))
}
}
\references{
\itemize{
\item{\href{https://geoservices.ign.fr/adminexpress#telechargementCog}{couches cartographiques ADMIN-EXPRESS-COG (IGN)}}}
}
