% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changement_COG_varNum.R
\encoding{UTF-8}
\name{loadMap}
\alias{loadMap}
\title{Charger la couche cartographique adaptée à vos données}
\usage{
loadMap(destfile = tempdir(), COG = annee_ref, nivsupra,
  donnees_insee = F)
}
\arguments{
\item{destfile}{indique le "path" où sera enregistrée la couche cartographique téléchargée (4 fichiers shp,shx,prj et dbf). Par défaut vaut tempfile() (répertoire temporaire)}

\item{COG}{indique l'année de COG de la table communale considérée. (exemple 2017). Années possibles : de 2015 à 2017. Par défaut, vaut annee_ref.}

\item{nivsupra}{est une chaîne de caractères qui indique le nom du niveau supra-communale souhaité. Plus précisément :
- "DEP" : départements
- "REG" : régions
- "EPCI" : EPCI au 01/01/20XX
- "ARR" : arrondissements au 01/01/20XX
- "CV" : cantons-villes au 01/01/20XX
- "ZE2010" : zones d'emploi 2010
- "UU2010" : unités urbaines 2010
- "AU2010" : aires urbaines 2010
- "BV2012" : bassins de vie 2012}

\item{donnees_insee}{vaut TRUE si les données manipulées sont produites par l'Insee. En effet, quelques rares modifications communales (la défusion des communes Loisey et Culey au 1er janvier 2014 par exemple) ont été prises en compte dans les bases de données communales de l'Insee plus tard que la date officielle.}
}
\description{
Charger la couche cartographique adaptée à vos données en indiquant l'année du code officiel géographique (COG) ainsi que le niveau géographique (communal ou supra-communal) souhaités
}
\details{
Le code officiel géographique le plus récent du package est actuellement celui au 01/01/2017. \cr

Les millésimes des COG qui peuvent être utilisés sont à ce stade les suivants : 2015, 2016 et 2017. \cr
}
\examples{
## Exemple 1
## Do not run (downloads files to your tempdir())
# reg_sf <- loadMap(COG=2016,nivsupra="REG")
# par(mar=c(0,0,0,0))
# plot(sf::st_geometry(reg_sf))
}
\references{
\itemize{
\item{\href{http://professionnels.ign.fr/adminexpress}{couches cartographiques Admin-Express (IGN)}}
\item{\href{http://professionnels.ign.fr/geofla}{couches cartographiques GEOFLA (IGN)}}}
}
