% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionner_DOM_glisser.R
\encoding{UTF-8}
\name{positionner_DOM_glisser}
\alias{positionner_DOM_glisser}
\title{Indiquer la position souhaitée des DOM en les glissant à partir d'une interface graphique}
\usage{
positionner_DOM_glisser(objet, nivsupra = colnames(objet)[1])
}
\arguments{
\item{objet}{est un fond de carte de la France (métropolitaine et 5 DOM) de type sf. Les niveaux géographiques possibles sont ceux du package CARTElette c'est-à-dire :
- "DEP" : départements
- "REG" : régions
- "EPCI" : EPCI au 01/01/20XX
- "ARR" : arrondissements au 01/01/20XX
- "CV" : cantons-villes au 01/01/20XX
- "ZE2010" : zones d'emploi 2010
- "UU2010" : unités urbaines 2010
- "AU2010" : aires urbaines 2010
- "BV2012" : bassins de vie 2012}

\item{nivsupra}{est une chaîne de caractères qui indique le nom du niveau supra-communal concerné. Cette chaîne de caractère doit également correspondre à la colonne de l'objet sf contenant les codes géographiques. Il peut s'agir de :
- "DEP" : départements
- "REG" : régions
- "EPCI" : EPCI au 01/01/20XX
- "ARR" : arrondissements au 01/01/20XX
- "CV" : cantons-villes au 01/01/20XX
- "ZE2010" : zones d'emploi 2010
- "UU2010" : unités urbaines 2010
- "AU2010" : aires urbaines 2010
- "BV2012" : bassins de vie 2012}
}
\description{
Indiquer la position souhaitée des DOM en les glissant à partir d'une interface graphique
}
\details{
La fonction renvoie une liste de coordonnées pour les futurs centroïdes
}
\examples{
\dontrun{
ze <- charger_carte(COG=2019,nivsupra="ZE2010")
positions <- positionner_DOM_glisser(objet=ze)
positions
}
}
\references{
\itemize{
\item{\href{http://professionnels.ign.fr/adminexpress}{couches cartographiques Admin-Express (IGN)}}
\item{\href{http://professionnels.ign.fr/geofla}{couches cartographiques GEOFLA (IGN)}}}
}
