% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionner_DOM_grille.R
\encoding{UTF-8}
\name{positionner_DOM_grille}
\alias{positionner_DOM_grille}
\title{Indiquer la position souhaitée des DOM à partir d'une grille géographique}
\usage{
positionner_DOM_grille(projection = NULL)
}
\arguments{
\item{projection}{projection de la couche cartographique à modifier. vaut NULL si la projection souhaitée est WGS84 ("+proj=longlat +datum=WGS84 +no_defs"). Sinon peut être devinée en faisant sf::st_crs(couche_sf)$proj4string}
}
\description{
Indiquer la position souhaitée des DOM à partir d'une grille géographique
}
\details{
La fonction renvoie une liste de coordonnées pour les futurs centroïdes
}
\examples{
\dontrun{
positions <- positionner_DOM_grille()
positions
}
}
\references{
\itemize{
\item{\href{http://professionnels.ign.fr/adminexpress}{couches cartographiques Admin-Express (IGN)}}
\item{\href{http://professionnels.ign.fr/geofla}{couches cartographiques GEOFLA (IGN)}}}
}
